+++ CHANGELOG.md
@@ -2,7 +2,18 @@
 
 
 ## Version 1.0
+### 1.0.2 - 2017-09-04
+- **Added**
+  - Ability to use backspace key
+  - Reads direct user input rather than through System.console()
+  - Storing commands in history (deprecated - to be fully implemented in later version)
+- **Removed**
+  - Support for taking user input on windows (runnable but untested) - Will be reimplemented in later version
+- **Issues**
+  - Bug when using arrow keys
+  - Bug when taking user input inside an IDE
+
 ### 1.0.1 - 2017-08-31
-#### Fixed
-- Bug involving zero-parameter usage
-- Extra lines in CLI
+- **Fixed**
+  - Bug involving zero-parameter usage
+  - Extra lines in CLI